最新

+++ common/datapacket/datapacket.h
@@ -246,7 +246,7 @@ struct sObjData
     sDataUnit vol; // 电压
     sDataUnit cur; // 电流
 
-    ushort pow[PACK_ARRAY_SIZE]; // 功率
+    uint pow[PACK_ARRAY_SIZE]; // 功率
     uint ele[PACK_ARRAY_SIZE]; // 电能
 
     uchar pf[PACK_ARRAY_SIZE]; // 功率因数
